Move utils to Kotlin out of K/N repo
