+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/ir/util/IrUtils2.kt
@@ -20,7 +20,6 @@ import org.jetbrains.kotlin.config.LanguageVersionSettings
 import org.jetbrains.kotlin.descriptors.*
 import org.jetbrains.kotlin.descriptors.FunctionDescriptor
 import org.jetbrains.kotlin.descriptors.ParameterDescriptor
-import org.jetbrains.kotlin.descriptors.ValueParameterDescriptor
 import org.jetbrains.kotlin.descriptors.annotations.AnnotationDescriptorImpl
 import org.jetbrains.kotlin.descriptors.annotations.Annotations
 import org.jetbrains.kotlin.ir.IrElement
@@ -89,219 +88,6 @@ internal fun IrFile.addTopLevelInitializer(expression: IrExpression, context: Ko
     }
     addChild(irField)
 }
-
-fun IrClass.addFakeOverrides(symbolTable: SymbolTable) {
-
-    val startOffset = this.startOffset
-    val endOffset = this.endOffset
-
-    descriptor.unsubstitutedMemberScope.getContributedDescriptors()
-            .filterIsInstance<CallableMemberDescriptor>()
-            .filter { it.kind == CallableMemberDescriptor.Kind.FAKE_OVERRIDE }
-            .forEach {
-                this.addChild(createFakeOverride(it, startOffset, endOffset, symbolTable))
-            }
-}
-
-private fun createFakeOverride(
-        descriptor: CallableMemberDescriptor,
-        startOffset: Int,
-        endOffset: Int,
-        symbolTable: SymbolTable
-): IrDeclaration {
-
-    fun FunctionDescriptor.createFunction(): IrSimpleFunction = IrFunctionImpl(
-            startOffset, endOffset,
-            IrDeclarationOrigin.FAKE_OVERRIDE,
-            this,
-            symbolTable.translateErased(this@createFunction.returnType!!)
-    ).apply {
-        createParameterDeclarations(symbolTable)
-    }
-
-    return when (descriptor) {
-        is FunctionDescriptor -> descriptor.createFunction()
-        is PropertyDescriptor ->
-            IrPropertyImpl(startOffset, endOffset, IrDeclarationOrigin.FAKE_OVERRIDE, descriptor).apply {
-                // TODO: add field if getter is missing?
-                getter = descriptor.getter?.createFunction()
-                setter = descriptor.setter?.createFunction()
-            }
-        else -> TODO(descriptor.toString())
-    }
-}
-
-fun IrFunction.createParameterDeclarations(symbolTable: SymbolTable) {
-
-    fun ParameterDescriptor.irValueParameter() = IrValueParameterImpl(
-            innerStartOffset(this), innerEndOffset(this),
-            IrDeclarationOrigin.DEFINED,
-            this, symbolTable.translateErased(this.type),
-            (this as? ValueParameterDescriptor)?.varargElementType?.let { symbolTable.translateErased(it) }
-    ).also {
-        it.parent = this@createParameterDeclarations
-    }
-
-    dispatchReceiverParameter = descriptor.dispatchReceiverParameter?.irValueParameter()
-    extensionReceiverParameter = descriptor.extensionReceiverParameter?.irValueParameter()
-
-    assert(valueParameters.isEmpty())
-    descriptor.valueParameters.mapTo(valueParameters) { it.irValueParameter() }
-
-    assert(typeParameters.isEmpty())
-    descriptor.typeParameters.mapTo(typeParameters) {
-        IrTypeParameterImpl(
-                innerStartOffset(it), innerEndOffset(it),
-                IrDeclarationOrigin.DEFINED,
-                it
-        ).also { typeParameter ->
-            typeParameter.parent = this
-            typeParameter.descriptor.upperBounds.mapTo(typeParameter.superTypes, symbolTable::translateErased)
-        }
-    }
-}
-
-fun IrSimpleFunction.setOverrides(symbolTable: ReferenceSymbolTable) {
-    assert(this.overriddenSymbols.isEmpty())
-
-    this.descriptor.overriddenDescriptors.mapTo(this.overriddenSymbols) {
-        symbolTable.referenceSimpleFunction(it.original)
-    }
-}
-
-fun IrClass.createParameterDeclarations(symbolTable: SymbolTable) {
-    this.descriptor.declaredTypeParameters.mapTo(this.typeParameters) {
-        IrTypeParameterImpl(startOffset, endOffset, IrDeclarationOrigin.DEFINED, it).apply {
-            it.upperBounds.mapTo(this.superTypes, symbolTable::translateErased)
-        }
-    }
-
-    this.thisReceiver = IrValueParameterImpl(
-            startOffset, endOffset,
-            IrDeclarationOrigin.DEFINED,
-            this.descriptor.thisAsReceiverParameter,
-            this.symbol.typeWith(this.typeParameters.map { it.defaultType }),
-            null
-    )
-}
-
-fun IrClass.setSuperSymbols(symbolTable: ReferenceSymbolTable) {
-    assert(this.superTypes.isEmpty())
-    this.descriptor.typeConstructor.supertypes.mapTo(this.superTypes) { symbolTable.translateErased(it) }
-
-    this.simpleFunctions().forEach {
-        it.setOverrides(symbolTable)
-    }
-}
-
-fun IrClass.simpleFunctions() = declarations.flatMap {
-    when (it) {
-        is IrSimpleFunction -> listOf(it)
-        is IrProperty -> listOfNotNull(it.getter, it.setter)
-        else -> emptyList()
-    }
-}
-
-fun IrClass.createParameterDeclarations() {
-    assert (thisReceiver == null)
-
-    thisReceiver = WrappedReceiverParameterDescriptor().let {
-        IrValueParameterImpl(
-                startOffset, endOffset,
-                IrDeclarationOrigin.INSTANCE_RECEIVER,
-                IrValueParameterSymbolImpl(it),
-                Name.special("<this>"),
-                0,
-                symbol.typeWith(typeParameters.map { it.defaultType }),
-                null,
-                false,
-                false
-        ).apply {
-            it.bind(this)
-            parent = this@createParameterDeclarations
-        }
-    }
-}
-
-fun IrFunction.createDispatchReceiverParameter(origin: IrDeclarationOrigin? = null) {
-    assert(dispatchReceiverParameter == null)
-
-    dispatchReceiverParameter = IrValueParameterImpl(
-            startOffset, endOffset,
-            origin ?: parentAsClass.origin,
-            IrValueParameterSymbolImpl(parentAsClass.thisReceiver!!.descriptor),
-            Name.special("<this>"),
-            0,
-            parentAsClass.defaultType,
-            null,
-            false,
-            false
-    ).apply {
-        parent = this@createDispatchReceiverParameter
-    }
-}
-
-fun IrClass.addFakeOverrides() {
-    fun IrDeclaration.toList() = when (this) {
-        is IrSimpleFunction -> listOf(this)
-        is IrProperty -> listOfNotNull(getter, setter)
-        else -> emptyList()
-    }
-
-    val overriddenFunctions = declarations
-            .flatMap { it.toList() }
-            .flatMap { it.overriddenSymbols.map { it.owner } }
-            .toSet()
-
-    val unoverriddenSuperFunctions = superTypes
-            .map { it.getClass()!! }
-            .flatMap { irClass ->
-                irClass.declarations
-                        .flatMap { it.toList() }
-                        .filter { it !in overriddenFunctions }
-            }
-            .toMutableSet()
-
-    // TODO: A dirty hack.
-    val groupedUnoverriddenSuperFunctions = unoverriddenSuperFunctions.groupBy { it.name.asString() + it.allParameters.size }
-
-    fun createFakeOverride(overriddenFunctions: List<IrSimpleFunction>) =
-            overriddenFunctions.first().let { irFunction ->
-                val descriptor = WrappedSimpleFunctionDescriptor()
-                IrFunctionImpl(
-                        UNDEFINED_OFFSET,
-                        UNDEFINED_OFFSET,
-                        IrDeclarationOrigin.FAKE_OVERRIDE,
-                        IrSimpleFunctionSymbolImpl(descriptor),
-                        irFunction.name,
-                        Visibilities.INHERITED,
-                        Modality.OPEN,
-                        irFunction.returnType,
-                        irFunction.isInline,
-                        irFunction.isExternal,
-                        irFunction.isTailrec,
-                        irFunction.isSuspend
-                ).apply {
-                    descriptor.bind(this)
-                    parent = this@addFakeOverrides
-                    overriddenSymbols += overriddenFunctions.map { it.symbol }
-                    copyParameterDeclarationsFrom(irFunction)
-                }
-            }
-
-    val fakeOverriddenFunctions = groupedUnoverriddenSuperFunctions
-            .asSequence()
-            .associate { it.value.first() to createFakeOverride(it.value) }
-            .toMutableMap()
-
-    declarations += fakeOverriddenFunctions.values
-}
-
-private fun IrElement.innerStartOffset(descriptor: DeclarationDescriptorWithSource): Int =
-        descriptor.startOffset ?: this.startOffset
-
-private fun IrElement.innerEndOffset(descriptor: DeclarationDescriptorWithSource): Int =
-        descriptor.endOffset ?: this.endOffset
 
 inline fun <reified T> stub(name: String): T {
     return Proxy.newProxyInstance(T::class.java.classLoader, arrayOf(T::class.java)) {
